Rename progress.py to task_cache.py
